requirements.txt
Django>=5.0,<7.0
psutil>=5.9.0
speedtest-cli>=2.1.3
python-dotenv>=1.0.0
gunicorn
whitenoise
dj-database-url
psycopg2-binary